% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tmt_plot.R
\name{generate_tmt_plot}
\alias{generate_tmt_plot}
\title{Generate a randomized TMT-A or TMT-B plot, mas já dá erro se N for grande demais}
\usage{
generate_tmt_plot(
  number_list,
  letter_list,
  min_distance = 1,
  element_size = 6,
  N = 13,
  type = "B",
  seed = NULL
)
}
\arguments{
\item{number_list}{A vector of numbers}

\item{letter_list}{A vector of letters}

\item{min_distance}{Minimum distance between elements}

\item{element_size}{Size of circles and text}

\item{N}{Number of elements (e.g. 13)}

\item{type}{"A" or "B" for TMT version}

\item{seed}{Optional seed for reproducibility}
}
\value{
A ggplot2 plot
}
\description{
Generate a randomized TMT-A or TMT-B plot, mas já dá erro se N for grande demais
}
