% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_letter_set.R
\name{get_letter_set}
\alias{get_letter_set}
\title{Get letters from different scripts}
\usage{
get_letter_set(script = "latin", N = 26)
}
\arguments{
\item{script}{Character: "latin", "greek", "arabic", or "cyrillic"}

\item{N}{Number of letters to return}
}
\value{
Character vector of letters
}
\description{
Get letters from different scripts
}
